Extract CollectionRenderer
